+++ src/main/java/memethespire/MemeTheSpire.java
@@ -2,10 +2,13 @@ package memethespire;
 
 import basemod.BaseMod;
 import basemod.interfaces.PostInitializeSubscriber;
+import basemod.interfaces.RelicGetSubscriber;
 import com.evacipated.cardcrawl.modthespire.lib.SpireInitializer;
+import com.megacrit.cardcrawl.relics.AbstractRelic;
+import memethespire.relicmemes.RelicModification;
 
 @SpireInitializer
-public class MemeTheSpire implements PostInitializeSubscriber {
+public class MemeTheSpire implements PostInitializeSubscriber, RelicGetSubscriber {
     public static void initialize() { new MemeTheSpire(); }
 
     public static Config config;
@@ -19,5 +22,10 @@ public class MemeTheSpire implements PostInitializeSubscriber {
         config = Config.loadConfig();
         Config.setupConfigMenu();
         MemeCollection.loadMemeCollections();
+    }
+
+    @Override
+    public void receiveRelicGet(AbstractRelic abstractRelic) {
+        RelicModification.restore(abstractRelic);
     }
 }